add web hook backend (#3)
This change adds a webhook api that uses type hints to register event listeners. Documentation will be added in a future PR and this package should be moved into a separate repo when it's ready. The type definitions for the Github webhook payloads are not complete, but should be enough for what we need in this project. In a later PR detail should be added to cover the full payloads.