feat: support running multiple kodiak apps on one repo (#35)
Enable specifying required github app ID in config via app_id field. This allows us to install the production kodiak app globally and configure the development app to run on one specific repo by adding the development app's github id to the config.

This change is only useful to the development of kodiak.