ref: reduce unnecessary pull request processing
By storing events in a set we can cut down on processing pull requests
multiple times in short succession.

Previously, events for a pull request could be entered multiple times
into the redis queue. Now a pull request can only be in the queue once.

We're using `bzpopmin`, which is only supported by redis v5 or later. It
is also not supported by asyncio_redis yet, so we need to make a change
there.