fix: incorrect PR review validity check (#134)
Check that a user has write permission to make a meaningful review.


### NOTES
We need to find if a reviewer has write permission to a repo. This doesn't appear possible through the GraphQL api as the collaborators attribute only returns users and not their permissions. The v3 api may work.

The [collaborators list endpoint](https://developer.github.com/v3/repos/collaborators/#list-collaborators) appears bugged and doesn't show all users with write permissions to a repo. I've opened a ticket with Github, but, like many of my tickets, I haven't heard anything.
I've also posted on the community forum: https://github.community/t5/GitHub-API-Development-and/How-can-I-view-users-with-write-permissions-to-repo-through-the/td-p/28190

I think we need to do a query per reviewer on the [collaborators detail endpoint](https://developer.github.com/v3/repos/collaborators/#review-a-users-permission-level) to see if they have write permission to the repo. We could cache this if necessary.


Fixes #109