Add ability to do smart updating without merging  (#187)
## Why

In our workflow we need the ability to update a PR without merging it. This is because Kodiak does not support signed commits and as such we would like to keep our branches up to date once they are mergeable but wait to be manual merged. 

## What 

I think you were really close to this feature already. 

The first thing is your `update_pr_immediately_if_configured` already nicely waits for the PR to mergeable before it starts updating:

`update_pr_immediately_if_configured` only triggers if two things are true:
	1) `update_branch_immediately` flag is `true` 
	2) `m_res == MergeabilityResponse.NEEDS_UPDATE`
`1)` will protect this function from running for individuals who do not want this option. And `2)` actually already waits for the "smart updating" cycle because `MergeabilityResponse` prioritizes branch protection checks over `NEEDS_UPDATE` which means the branch will only be flagged as needing updating once it is ready to merge.

So the only thing I added was a new variable `do_not_merge` which will ensure that we always short circuit this function and do not add it to the merging queue. 

This will allow you to combine the flags `do_not_merge` and `update_pr_immediately_if_configured` to use kodiak to do smart auto updating in your PRs without merging. 

## Notes

If this idea looks like it's along the right path I can add more documentation/information to the README.

It looks like there are some other places Im missing the config variable. I can add those in as well shortly.