ref: GitHub StatusEvent parsing to be more lenient around URLs
Previously we were using `pydantic`'s `UrlStr` which includes a regex
check that was failing in some cases.

We don't store the events so I'm not entirely sure the value returned
by GitHub so I copied and updated a fixture to produce a similar error
to the one in Sentry.

SENTRY: KODIAK-52