ref: mergeability logic to be more centralized and extensible (#207)
As a result of Kodiak's development, application logic is sprinkled in many places throughout the app. This makes it hard to reason about Kodiak's behavior and add new features.

This change centralizes Kodiak's logic in one place to make it easier to modify Kodiak (for #198 merge.update_ignore_checks). This new setup is also more testable.

CHANGES
- move application logic to mergeable function. This function uses an API protocol that we sub in for testing. This API protocol should be kept basic, with higher level logic stored in mergeable.
- rewrite all tests for evaluation logic. The original tests were simple, but they weren't as useful as the current system because they only tested if exceptions were thrown, not what happens afterwards.
- remove pytest-timeout because it wasn't working with asyncio for the regex DoS test.
- refactor PullRequest class to move API calls to queries.py and satisfy API protocol for mergeable function.

I think we will be able to simplify some of the event processing logic with this new system.