fix: handling of rebase merge conflicts
Merge conflicts are not the same as rebase merge conflicts. It's possible for a PR to be mergeable for one method (like merge or squash) but for rebase to have it's own conflict. Now we check the canBeRebased property on pull requests to determine if there is a rebase merge conflict. We only check when the configured merge method is rebase.

fixes #222