wip: add support for bot reviews
a219e60
The collaborators API we use to check permissions on GitHub users does not work with bots, so we were getting API errors when a bot reviewed a PR. This change adds support for bot reviews.

If a bot reviews a PR, it means they have write access to a PR, so they're review "counts" as a review of a user with write access to a repository.

TODO
- [ ] automated tests if possible
- [ ] manual testing