fix: delete_branch_on_merge to ignore branches with dependencies
Currently we delete a branch if the PR is merged and `merge.delete_branch_on_merge` is configured. This behavior ignores the fact that multiple PRs can be dependent on a branch, so by deleting a branch we end up closing the PRs that are also depending on that branch.

Example:

A PR alpha against master from base-branch, a PR bravo against base-branch from leaf-branch.

Kodiak will delete base-branch when PR alpha is merged, causing PR bravo to also close because of the dependency on base-branch.

Now Kodiak will _not_ delete base-branch when PR bravo exists.