Revert rebase merge conflicts handling (#225)
This reverts commit 9ebe7ae56b44a6604c36cded29d5afc2c3a5d85f.

When I quickly opened a PR during testing, Kodiak raised a merge conflict saying the PR couldn't be rebased, but this went away after Kodiak left a comment. I believe GitHub was returning false for canBeRebased when the PR was just opened. This triggered Kodiak to believe the PR couldn't be rebased, when actually GitHub simply didn't know if the PR couldn't be rebased.

I'm not sure how to handle this, but I'm more concerned about raising false positives than the likelihood of rebase conflict without a general merge conflict.