revert invalid fix for delete_branch_on_merge (#232) (#237)
This reverts commit 935c3820dd5d8ebe387ecdd68569c2cea51eca5e.

From testing on staging it appears that this change did not fix the problem of deleting branches that had dependencies. headRef->associatedPullRequests does not return all the PRs associated with a ref, so this change is broken.

I'm not sure how we'll be able to fix this using the GitHub api.