fix: bug with status event logic
When we receive a status webhook event we look at the commit information to determine which PRs are associated with this event. Our API call GitHub incorrectly used the `head` query param by passing a SHA when it wanted a "org_name:ref_name". This error meant that our API call was returning _all_ open PRs instead of just the ones that depended on the branch.

TODO
- get PR with base of SHAs and PR with head of SHA