fix: bug with status event logic (#248)
When we receive a status webhook event we look at the commit information to determine which PRs are associated with this event.

Previously our API call to GitHub incorrectly used the `head` query param by passing a SHA when it wanted a "org_name:ref_name". This error meant that our API call was returning _all_ open PRs instead of just the ones that depended on the branch.

Now we only fetch all open PRs in case we cannot find references for the current commit. This covers the case of PRs opened from forks, which the status event doesn't provide an owner for to make a GitHub API query.