fix(web-api): authentication issues (#282)
Update authentication flow to work with web ui. I want the app and the api to have different hosting, so we won't have NGINX to do proxying. This means we have to fight around CORs, so the authentication process is a little tricky (at least to me).

I have an urge to use token authentication, but this session stuff appears to work for now.