add(web_ui): redirect for authed url
If a user attempts to access an authenticated page but is not logged in, they are redirected to the login page and are redirected to the root page upon successful authentication.

This change will result in the user being redirected to the page they attempted to access but were missing authentication. For example, if a user attempts to access https://app.kodiakhq.com/t/f8c69022-fcba-42a5-9e85-da3a466e6008/ without authentication, after they login successfully, the will be redirected back to that page!