add(web): management commands for aggregations and docker for web_ui (#298)
To have a staging environment we cannot use Netlify for the react app. This change adds a docker container to serve our React app instead.
I've also moved all of the aggregations into django management commands because I was having we're pathing issues on our servers and this was a clean fix.

CHANGES
- move web_ui from Netlify to docker container
- move aggregations into django managements commands
- handle database conflicts for user_pull_request_activity
- add graceful termination for event ingestion
- configure Sentry for backend and add endpoint to verify Sentry is working
- configure logging for django
- add logic to determine web api domain from app hostname

This code is the currently deployed app.

Co-authored-by: Christopher Dignam <1929960+chdsbd@users.noreply.github.com>