add: channel for triggering bot to reevaluate PRs
When a user hits the paywall, they will need to start a trial or subscription to remove it. However, Kodiak only evaluates a PR when it gets a new related GitHub webhook. So after starting a trial or subscription the user would still see paywalls on their PRs unless they trigger a refresh by updating the PR.

This change adds a method for triggering the bot to reevaluate PRs outside of GitHub webhook triggers. This change adds a process that will listen for events in Redis and trigger updates in the bot.

On trial or subscription upgrades the web api will send a refresh message via Redis to the listening process. The listening process will enqueue webhook events for all open PRs into the webhook event queue, like a normal GitHub webhook event, triggering reevaluation of PRs.