fix(web_ui): start subscription form seat selection clarity
Previously the start subscription form defaulted to 1 seat. This is a bad default and has caused issues for some customers. We now default to the active user count.

We also change the "Select at least N seats to continue service." font color to red when the user has not selected enough seats.