fix(web_ui): start subscription form seat selection clarity (#367)
Previously the start subscription form defaulted to 1 seat. This is a bad default and has caused issues for some customers. We now default to the active user count.

We also change the "Select at least N seats to continue service." font color to red when the user has not selected enough seats.

**default state**
<img width="433" alt="Screen Shot 2020-05-08 at 11 44 35 PM" src="https://user-images.githubusercontent.com/1929960/81463707-984dbd00-9189-11ea-9831-f632fc489054.png">

**too few seats**
<img width="443" alt="Screen Shot 2020-05-09 at 12 09 03 AM" src="https://user-images.githubusercontent.com/1929960/81463720-a3a0e880-9189-11ea-9250-8aeb869a3b3e.png">