add(bot): prioritize merge conflict notifications (#371)
fixes https://github.com/chdsbd/kodiak/issues/276

This problem is #276 is that the `merge.blacklist_title_regex` was taking priority over `merge.notify_on_conflict`. This is not as desirable because I think merge conflict notifications are higher priority.

I was going to make a more general change in #364, but I think this is a simple solution that addresses the immediate problem. I plan to circle back to #364 later.