fix(web_api): oauth flow not handling api error in membership check
Calling GET https://api.github.com/orgs/:org/memberships/:user was failing. I'm not sure yet why this was failing, but as a quick fix to make sure the login flow always works, I'm making this change.

My suspicion is that if the user is a collaborator instead of a member or admin of an organization, they will get a 403 for this API request.