ref(bot): simplify evaluation testing to be more readable (#383)
We have a lot of unnecessary information in these evaluation.py tests. This change makes it so we only have information that's important to the test.

We're basically replacing all the pytest fixtures with plain functions.

I've changed one test as an example. I plan to make more changes in chunks via other PRs.