fix(bot): queue on unknown mergeability status (#380)
When we get an UNKNOWN mergeability status we attempt to trigger a test commit by calling GET on the pull request. Internally GitHub will trigger an evaluation of the PR mergeability which should resolve the UNKNOWN merge ability status check.

My concern is that we don't always get a new webhook event when we call this endpoint. To prevent this we now queue the webhook event again explicitly.

I also noticed a couple issues with returns breaking our merge poll. I wonder if there is a way to prevent this.

related: #261