fix(bot): ignore conflicts when the PR is closed or merged
I encountered this weird behavior where after merge Kodiak said their was a merge conflict. I believe this regression was introduced in https://github.com/chdsbd/kodiak/pull/371, because we would return early if the PR was merged or closed, avoiding the merge conflict logic.

I think the source of the issue may be that we're getting a MergeStateStatus.DIRTY result upon merge. I wonder if we can just rely on MergeableState.CONFLICTING instead.

As a more reliable fix, I've added a check so we only leave a comment of the pull request is open.