fix(bot): ignore conflicts when the PR is closed or merged (#391)
Post merge Kodiak said their was a merge conflict. This problem was introduced in #371, because prior to that we would return early if the PR was merged or closed, avoiding the merge conflict logic.

When a pull request is merged we will get `MergeStateStatus.DIRTY` and `MergeableState.CONFLICTING `, which would normally trigger a merge conflict error. However it's not helpful to leave a comment when the pull request is merged, so now we check and only comment if the pull request is open.