wip: fix(bot): retrying on GitHub API internal server error
It is dangerous to retry merging when GitHub returns an internal server error (HTTP status code 500) because the merge can partially succeed in a way that will allow Kodiak to erroneously merge in a pull request multiple times.

The GitHub pull request merge API is _not always_ idempotent (https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button). Normally if you try to merge an already merged pull request using this API you'll get a `405 Method not allowed` error, as documented in the GitHub API docs. However in an error state, when the GitHub API returns a 500 HTTP error, the pull request can be partially merged, as shown in #397.

It appears the git operations of merging the pull request branch into master can succeed but not affect the pull request state itself. In this failure scenario the pull request will still be in an OPEN state even though the underlying branch was merged successfully into the target branch.

This change disables Kodiak merging a pull request when it encounters this scenario. When `merge.require_automerge_label` is enabled, Kodiak will remove the label and add a comment to the pull request explaining the problem. If `merge.require_automerge_label` is disabled Kodiak adds a label `merge.merge_failure_label` and leaves a similar comment. Kodiak will not merge a pull request when the automerge label is missing and `merge.require_automerge_label` is enabled or `merge.merge_failure_label` is applied.
The user must add/remove these labels to re-enable Kodiak merging the pull request.

fixes #397