fix(bot): disable Kodiak on GitHub API merge internal server error (#398)
This change disables Kodiak merging a pull request when it encounters an internal server error from the GitHub merge api. In this event, Kodiak adds a label `disable_kodiak_label` and leaves a comment describing the issue. Kodiak will not update, approve, comment on, label, or merge a pull request with the `disable_kodiak_label`. It will still update status checks. The user must remove this label to re-enable Kodiak merging the pull request.

fixes #397

## background
It is dangerous to retry merging when GitHub returns an internal server error (HTTP status code 500) because the merge can partially succeed in a way that will allow Kodiak to erroneously merge in a pull request multiple times, as we've seen in #397.

The GitHub pull request merge API is _not always_ idempotent (https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button). Normally if you try to merge an already merged pull request using this API you'll get a `405 Method not allowed` error, as documented in the GitHub API docs. However in an error state, when the GitHub API returns a 500 HTTP error, the pull request can be partially merged, as shown in #397.

It appears the git operations of merging the pull request branch into master can succeed but not affect the pull request state itself. In this failure scenario the pull request will still be in an OPEN state even though the underlying branch was merged successfully into the target branch.


## further discussion
I realized implementing this change that GitHub API calls to add pull request labels could also fail. In this case Kodiak would still erroneously merge a pull request.

I think the most robust solution to this GitHub API reliability problem would be to store some information in Redis to disable Kodiak. But then there's the problem of how to surface that data in the GitHub UI. If we can't trust the GitHub API, how can we notify a user that Kodiak is disabled on their pull request? 🙍‍♂️ 

## screenshots
<img width="660" alt="image" src="https://user-images.githubusercontent.com/1929960/82390717-7831bf80-9a0d-11ea-86ff-ab3b288ef370.png">