fix(bot): escaping in delete branch and delete label api calls (#408)
If a branch had a name that was not url safe, like `#123`, the resulting url we would call out to GitHub with would return 404 because `#` would leave the URL cut off.

The fix is to url encode the branch name. I've checked other queries and found one suspicious one and a false alarm.

If we pass url-unsafe parameters via `params` requests will escape them properly:

https://github.com/chdsbd/kodiak/blob/071ece2258de3eaa7dd34ec2db661d3c8ab3fc0b/bot/kodiak/queries.py#L902-L907

I think if a label was unsafely named we would fail to delete it. So I fixed this api call:

https://github.com/chdsbd/kodiak/blob/071ece2258de3eaa7dd34ec2db661d3c8ab3fc0b/bot/kodiak/queries.py#L1002-L1006

Fixes #407