add: merge.message.include_coauthors (#420)
Add configuration option `merge.message.include_coauthors` to enable Kodiak to include coauthors in pull request body.
This option only applies to pull requests with `merge.message.body = "pull_request_body"`.

We add coauthors based on the last 100 commits on a pull request. We also only add coauthor information for commits that are linked to a GitHub user. We use the GitHub user's account information (login, id, name if available) to generate the coauthor trailers, not the commit author name or email.

resolves #277