ref(web): move app and api to same domain to combat CORS (#429)
Having the React app and the Django API on separate domains has worked, but I recently wanted to send a JSON payload to the API and this is not possible with CORS. The Content-Type `application/json` is not an allowed Content-Type for cross origin requests and session cookies do not get sent when this Content-Type is set.

Moving the API to the same domain as the app resolves these CORS issues.

This PR removes the logic for sending api requests to a different domain. To support older web ui bundles the backend logic will be removed in another PR and deployed later.

For deployment we use HAProxy to handle routing api requests to the appropriate service.