fix(bot): error handling to be less noisy (#436)
It appears that "user" can frequently be null. This is fine because we handle the errors, but we also log a warning which is noisy in Sentry.

We explicitly handle the null case so we don't log a warning about it.

This test doesn't really do much but I manually verified that we don't log a warning with this case.