add(bot): use valid repo merge method if not specified (#464)
If `merge.method` is not configured, use the first valid merge method in the repository based on the list "merge", "squash", "rebase".

For example, if a repository only allows "rebase" commits and `merge.method` is not provided, Kodiak will now use the "rebase" merge method. Previously it would have used the default "merge" method.

If a repository only allowed "squash" and "rebase", and `merge.method` was not configured, "squash" would be chosen because it is the first valid option in the list "merge", "squash", "rebase".

This change is backwards compatible for the most part. Kodiak now works instead of raising a configuration error if a user only allowed "squash" merge methods and didn't specify `merge.method`. If a user was relying on `"merge"` being the default, they will be okay since `"merge"` is the first option in the list of valid merge methods to try.

Fixes https://github.com/chdsbd/kodiak/issues/387

Docs will be updated in https://github.com/chdsbd/kodiak/pull/466.