update(bot): evaluation timeout to reduce errors (#472)
I think it was a mistake to not handle these timeout errors. If we hit a timeout Kodiak's basically dead until it receives another message for that pull request.

I think a better solution is to queue the pull request again on timeout. This way we won't loose the pull request and it won't block others because it will go at the end of the queue.

I'm also bumping the timeout because with our api rate limit we've been hitting a lot of timeouts recently.

FYI, we don't have any test coverage on this code.

_This pull request is best viewed with whitespace changes hidden_