add(web): better subscription exemption support
This change adds an easier way to mark an account as exempt from subscriptions.

GitHub Sponsors, Enterprise Plan customers, non profits, etc, do not require subscriptions. The current solution is to set an trial that expires in 10 years, but this is a hack.

This change adds a flag we can set in the database to exempt an account and copy that we can set to provide a custom message.