fix: `automerge_labels` => `automerge_label` (#590)
## Issue

When implementing this **awesome** tool 🙏, I observed that `automerge_label` accepts a `string` or `array`. 
I attempted to use  `automerge_labels`, which matches the docs but causes an error when Github Checks ran. 

> I believe the error was, **"automerge_label is not defined"**.

## Fix

I continued to use my` automerge_label` **array** to fix the issue but changed the text from `automerge_labels` to `automerge_label`

Thanks for your great work!