add(bot): rebase fast forward merge
For GitHub's [merge pull request API](https://docs.github.com/en/rest/reference/pulls#merge-a-pull-request), rebased commits will be recreated on merge. This is counter to normal git rebase. To preserve the commits on merge we can use [ref API](https://docs.github.com/en/rest/reference/git#update-a-reference) to point the base branch to the new commits. This way the commits in the pull request are merged unchanged.