fix(bot): merge queue merging timestamp
With the recent changes to the merge queue display, the merging pull requests looses the time it was added to the queue and instead shows the time it started merging.

We can use the sorted set score to track the time the pull request was added to the merge queue and display that time when merging.