ref(bot): incoming webhook processing
Right now the incoming webhook HTTP server is kind of mixed in with the
event processing code.

This is an issue when it comes to deploy code because we can't stop only
the event processing code, instead we stop the HTTP as well which causes
us to drop events.

New plan, we'll have a couple entrypoints:

- HTTP server handling incoming webhooks which does as little processing
  as possible before dumping the events onto a queue
- Queue consumers, the existing logic would work off the new incoming
  event queue instead of in the HTTP server's endpoint handler.